Add README entry about changing the secret token.
Without changing this token, users of this sample
application will be susceptible to session-modification
attacks, because anyone could use the current secret token
to sign modified session data.

Additionally, we may want to add a check in an initializer
that verifies that the secret token has been changed when in the
production environment, and if not, displays a warning message or
halts the application.